[Gold II] Title: 꼬인 전깃줄, Time: 12 ms, Memory: 2924 KB -BaekjoonHub

+++ 백준/Gold/1365. 꼬인 전깃줄/꼬인 전깃줄.cc
@@ -8,7 +8,7 @@ int lis(vector<int>&arr) {
         if (it==lis.end()) lis.push_back(x);
         else *it=x;
     }
-    return static_cast<int>(lis.size());
+    return lis.size();
 }
 
 int main() {
@@ -16,6 +16,7 @@ int main() {
 
     int n;
     cin>>n;
+
     vector<int>arr(n);
     for (int i=0; i<n; i++) {
         cin>>arr[i];
